refactor: Decouple listener.Cleanup from tun
The listener.Cleanup method will be called during
executor.Shutdown and route.restart, so it should serve
the entire listener rather than a single tun device.

Currently listener.ReCreateTun will call it to handle
some internal affairs, but this is unnecessary.

At the same time, it is also not necessary to manually
close the tunLister during Shutdown and restart.

But the listner.Cleanup method is still worth keeping
in order to deal with some additional cleanup tasks
outside the process life cycle for other types of
listeners in the future.